pyproject.toml
[project]
name = "cvops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fire>=0.7.1",
    "openai>=2.15.0",
    "pydantic-ai>=1.42.0",
    "python-dotenv>=1.2.1",
    "rendercv[full]>=2.2",
]
